about to bail on locations
tried a few things.  The aframe scene inspector is super cool but didn't find much, and I went back to the working codepen example from the docs - didn't work in either safari or chrome on the roof.  Safari had the text hosted on the camera, chrome showed nothing.

Maybe a few more tries but might have to stick with markers for the weekend